chg: moved FilterEnumerator class to separate script file #3
