+++ Assets/Scripts/Ecs/EcsFilter.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
@@ -22,43 +21,4 @@ public class EcsFilter : IEnumerable<EcsEntity>
     {
         return GetEnumerator();
     }
-}
-
-public class FilterEnumerator : IEnumerator<EcsEntity>
-{
-    private readonly List<EcsEntity> _entities;
-    private int _position;
-
-    public FilterEnumerator(IEnumerable<EcsEntity> entities)
-    {
-        _entities = new List<EcsEntity>(entities);
-        _position = -1;
-    }
-
-    public bool MoveNext()
-    {
-        _position++;
-        return _position < _entities.Count;
-    }
-
-    public void Reset() => _position = -1;
-
-    public EcsEntity Current
-    {
-        get
-        {
-            try
-            {
-                return _entities[_position];
-            }
-            catch (IndexOutOfRangeException)
-            {
-                throw new InvalidOperationException("Index out of range");
-            }
-        }
-    }
-
-    object IEnumerator.Current => Current;
-
-    public void Dispose() => _entities.Clear();
 }
