+++ Assets/Scripts/Ecs/FilterEnumerator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0348321377677484b932a618a53f33e6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
